tests: Added webhook send tests
