--- src/index.js
@@ -1,5 +1,0 @@
-const { run } = require("./run.js");
-
-run()
-  .then(() => process.exit(0))
-  .catch((err) => core.setFailed(err.message));
