+++ src/webhooks.js
@@ -4,61 +4,33 @@ const MAX_MESSAGE_LENGTH = 128;
 module.exports.send = (
   id,
   token,
-  repo,
-  branch,
-  url,
-  commits,
-  size,
+  payload,
   hideLinks,
   censorUsername,
   color
-) =>
-  new Promise((resolve, reject) => {
-    let client;
-    console.log("Preparing Webhook...");
-    try {
-      client = new discord.WebhookClient(id, token);
-    } catch (error) {
-      reject(error.message);
-      return;
-    }
+) => {
+  const repository = payload.repository.full_name;
+  const commits = payload.commits;
+  const size = commits.length;
+  const branch = payload.ref.split("/")[payload.ref.split("/").length - 1];
+  const url = payload.compare;
 
-    client
-      .send(
-        createEmbed(
-          repo,
-          branch,
-          url,
-          commits,
-          size,
-          hideLinks,
-          censorUsername,
-          color
-        )
-      )
-      .then(() => {
-        console.log("Successfully sent the message!");
-        resolve();
-      }, reject);
-  });
+  if (commits.length === 0) {
+    // This was likely a "--tags" push.
+    console.log(`Aborting analysis, found no commits.`);
+    return;
+  }
 
-function createEmbed(
-  repo,
-  branch,
-  url,
-  commits,
-  size,
-  hideLinks,
-  censorUsername,
-  color
-) {
+  console.log(`Received payload: ${JSON.stringify(payload, null, 2)}`);
+  console.log(`Received ${commits.length}/${size} commits...`);
   console.log("Constructing Embed...");
+
   let latest = commits[0];
   const count = size == 1 ? "Commit" : " Commits";
 
   let embed = new discord.MessageEmbed()
     .setColor(color)
-    .setTitle(`⚡ ${size} ${count} - \`${repo}\` on 🌳 \`${branch}\``)
+    .setTitle(`⚡ ${size} ${count} - \`${repository}\` on 🌳 \`${branch}\``)
     .setDescription(getChangeLog(commits, size, hideLinks, censorUsername))
     .setTimestamp(Date.parse(latest.timestamp));
 
@@ -66,8 +38,24 @@ function createEmbed(
     embed.setURL(url);
   }
 
-  return embed;
-}
+  return new Promise((resolve, reject) => {
+    let client;
+    console.log("Preparing Discord webhook client...");
+    try {
+      client = new discord.WebhookClient(id, token);
+    } catch (error) {
+      reject(error.message);
+    }
+
+    return client
+      .send(embed)
+      .then((result) => {
+        console.log("Successfully sent the message!");
+        resolve(result);
+      })
+      .catch((error) => reject(error));
+  });
+};
 
 function getChangeLog(commits, size, hideLinks, censorUsername) {
   console.log("Constructing Changelog...");
